+++ tests/Unit/Module/Service/ModuleActivationsServiceTest.php
@@ -12,8 +12,11 @@ namespace OxidEsales\GraphQL\ConfigurationAccess\Tests\Unit\Module\Service;
 use OxidEsales\EshopCommunity\Internal\Transition\Utility\ContextInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Bridge\ModuleActivationBridgeInterface;
 use OxidEsales\GraphQL\ConfigurationAccess\Module\Exception\ModuleActivationException;
+use OxidEsales\GraphQL\ConfigurationAccess\Module\Exception\ModuleActivationBlockedException;
+use OxidEsales\GraphQL\ConfigurationAccess\Module\Exception\ModuleDeactivationBlockedException;
 use OxidEsales\GraphQL\ConfigurationAccess\Module\Exception\ModuleDeactivationException;
 use OxidEsales\GraphQL\ConfigurationAccess\Module\Service\ModuleActivationService;
+use OxidEsales\GraphQL\ConfigurationAccess\Module\Service\ModuleBlocklistServiceInterface;
 use OxidEsales\GraphQL\ConfigurationAccess\Tests\Unit\UnitTestCase;
 
 /**
@@ -24,7 +27,7 @@ class ModuleActivationsServiceTest extends UnitTestCase
     /**
      * @dataProvider activationDataProvider
      */
-    public function testModuleActivationAndDeactivation(
+    public function testModuleActivationAndDeactivationSuccess(
         string $method,
     ): void {
         $shopId = 1;
@@ -33,6 +36,12 @@ class ModuleActivationsServiceTest extends UnitTestCase
         $moduleActivationBridgeMock
             ->method($method)
             ->with($moduleId, $shopId);
+
+        $moduleBlocklistServiceMock = $this->createMock(ModuleBlocklistServiceInterface::class);
+        $moduleBlocklistServiceMock
+            ->method('isModuleBlocked')
+            ->with($moduleId)
+            ->willReturn(false);
 
         $sut = $this->getSut(
             context: $this->getContextMock($shopId),
@@ -47,7 +56,7 @@ class ModuleActivationsServiceTest extends UnitTestCase
     /**
      * @dataProvider exceptionDataProvider
      */
-    public function testModuleActivationAndDeactivationExceptions(
+    public function testModuleActivationAndDeactivationThrowsExceptions(
         string $method,
         mixed $exceptionClass
     ): void {
@@ -55,15 +64,54 @@ class ModuleActivationsServiceTest extends UnitTestCase
         $moduleActivationBridgeMock = $this->createMock(ModuleActivationBridgeInterface::class);
         $moduleActivationBridgeMock
             ->method($method)
+            ->with($moduleId, 1)
             ->willThrowException(new \Exception());
 
         $this->expectException($exceptionClass);
 
         $sut = $this->getSut(
+            $this->getContextMock(),
             moduleActivationBridge: $moduleActivationBridgeMock
         );
 
         ($method === 'activate') ? $sut->activateModule($moduleId) : $sut->deactivateModule($moduleId);
+    }
+
+    /**
+     * @dataProvider moduleBlockedExceptionDataProvider
+     */
+    public function testModuleActivationAndDeactivationBlockedException(
+        string $method,
+        mixed $exceptionClass
+    ) {
+        $moduleId = uniqid();
+        $moduleBlockListServiceMock = $this->createMock(ModuleBlocklistServiceInterface::class);
+        $moduleBlockListServiceMock
+            ->method('isModuleBlocked')
+            ->with($moduleId)
+            ->willReturn(true);
+
+        $sut = $this->getSut(
+            moduleBlocklistService: $moduleBlockListServiceMock
+        );
+
+        $this->expectException($exceptionClass);
+        ($method === 'activate') ? $sut->activateModule($moduleId) : $sut->deactivateModule($moduleId);
+    }
+
+    public static function moduleBlockedExceptionDataProvider(): \Generator
+    {
+        yield 'test activate module blocked exception' => [
+            'method' => 'activate',
+            'exceptionClass' => ModuleActivationBlockedException::class
+
+        ];
+
+        yield 'test deactivate module blocked exception' => [
+            'method' => 'deactivate',
+            'exceptionClass' => ModuleDeactivationBlockedException::class
+
+        ];
     }
 
     public static function activationDataProvider(): \Generator
@@ -94,13 +142,16 @@ class ModuleActivationsServiceTest extends UnitTestCase
 
     public function getSut(
         ContextInterface $context = null,
-        ModuleActivationBridgeInterface $moduleActivationBridge = null
+        ModuleActivationBridgeInterface $moduleActivationBridge = null,
+        ModuleBlocklistServiceInterface $moduleBlocklistService = null
     ): ModuleActivationService {
         return new ModuleActivationService(
             context: $context
                 ??  $this->createStub(ContextInterface::class),
             moduleActivationBridge: $moduleActivationBridge
-                ??  $this->createStub(ModuleActivationBridgeInterface::class)
+                ??  $this->createStub(ModuleActivationBridgeInterface::class),
+            moduleBlocklistService: $moduleBlocklistService
+                ??  $this->createStub(ModuleBlocklistServiceInterface::class)
         );
     }
 }