+++ CHANGELOG.md
@@ -10,17 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Theme list and filtering option on basis of theme name and status
 - Module list and filtering option on basis of module name and status
 - Activation of given theme by themeId
-- Mutations to activate and deactive a module.
+- Mutations to de/activate a module.
+- Prevention of de/activation of certain modules mentioned in modules_blocklist.yaml.
 
 ## [1.1.0] - 2024-07-05
 This is stable release for v1.1.0. No changes have been made since v1.1.0-rc.1.
 
 ## [1.1.0-rc.1] - 2024-05-30
-[.gitignore](.gitignore)
 ### Added
 - PHP 8.2 support
 - Module activation dependency on GraphQL Base module
--  Activate a given  theme by themeId.
 
 ### Changed
 - PHPUnit upgraded to version 10.x
@@ -33,6 +32,7 @@ This is stable release for v1.1.0. No changes have been made since v1.1.0-rc.1.
 
 - Initial release
 
+[1.2.0]: https://github.com/OXID-eSales/graphql-configuration-access/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/OXID-eSales/graphql-configuration-access/compare/v1.1.0-rc.1...v1.1.0
 [1.1.0-rc.1]: https://github.com/OXID-eSales/graphql-configuration-access/compare/v1.0.0...v1.1.0-rc.1
 [1.0.0]: https://github.com/OXID-eSales/graphql-configuration-access/releases/tag/v1.0.0